% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_funs.R
\name{amend.and.simulate.resamples}
\alias{amend.and.simulate.resamples}
\title{Convert to simulated resamples}
\usage{
amend.and.simulate.resamples(rset_obj)
}
\arguments{
\item{rset_obj}{an rset object created by make.rolling.custom.resamples}
}
\description{
This function takes an reset object and converts all OOS sets to simulated
data by way of MASS::mvrnorm.
}
