% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_iron_butterfly.R
\name{attach.option.returns}
\alias{attach.option.returns}
\title{Generate shifly returns}
\usage{
attach.option.returns(out)
}
\arguments{
\item{out}{an object created from make.shifly.values}
}
\description{
This function attaches the rolling options returns for the shifly values.
We use a linear method instead of log returns because we are aggregating across
strategies and not time.
}
