% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels_wrappers.R
\name{build.workflowset}
\alias{build.workflowset}
\title{Builds custom workflowset}
\usage{
build.workflowset(split)
}
\arguments{
\item{split}{a split created by running first.split once and then running it
again on the listed analysis set of the resulting split object.}
}
\description{
This function builds a workflowset that will be plugged into the
workflow_map function from workflowsets
}
