% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL_stock.R
\name{clean.symbols}
\alias{clean.symbols}
\title{Cleaning the ticker set}
\usage{
clean.symbols(feat_syms, my_sym)
}
\arguments{
\item{feat_syms}{a list of tickers to be used a features}

\item{my_sym}{a single ticker}
}
\description{
This function cleans the ticker list provided by removing your target
ticker from the set if and only if it is present inside the list.
}
