% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_iron_butterfly.R
\name{.calc.and.format.greeks}
\alias{.calc.and.format.greeks}
\title{Calculate Option Stuff - Internal}
\usage{
.calc.and.format.greeks(opt_args)
}
\arguments{
\item{opt_args}{an object created from the internal calls of
get.and.price.rolling.options}
}
\description{
This function uses the RQuantlib library to price options
}
