% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels_wrappers.R
\name{extract_best.oos.workflow}
\alias{extract_best.oos.workflow}
\title{Extract the best workflow}
\usage{
extract_best.oos.workflow(tuned_models, split, metric = "rmse")
}
\arguments{
\item{tuned_models}{a table created by tune.workflowsets.random.search}

\item{split}{a split object, preferably a validation split}

\item{metric}{a metric, class == character, that was supplied to the yardstick::metric_set in
tune.workflowsets.random.search}
}
\description{
This function helps extract the best performing workflow from the output of
tune.workflowsets.random.search
}
