% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels_wrappers.R
\name{first_split}
\alias{first_split}
\title{Short function to split the data}
\usage{
first_split(my_datasets, prop)
}
\arguments{
\item{my_datasets}{data sets created from yf.load.and.curate.financial.dataset}

\item{prop}{a fraction or a decimal with value between 0-1}
}
\description{
This function is a wrapper for initial_time_split from rsample
}
