% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL_stock.R
\name{gather_and.pad.yf.adj.price.data}
\alias{gather_and.pad.yf.adj.price.data}
\title{Clean the download Yahoo Finance data}
\usage{
gather_and.pad.yf.adj.price.data(symbols, yearz)
}
\arguments{
\item{symbols}{with the target ticker in first position,
supply this vector of tickers to retrieve adjusted close}

\item{yearz}{a look-back in amount of years, can be double or integer}
}
\description{
This function downloads data, reductively joins, by "index", each stock's
adj close vector, then pads the adj close using an updown method.
}
