% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL_stock.R
\name{generate.log.returns.data}
\alias{generate.log.returns.data}
\title{Create a log returns data set}
\usage{
generate.log.returns.data(price_data)
}
\arguments{
\item{price_data}{a data frame with adjusted close data and an index}
}
\description{
This function creates the "approximate invariants" from the
price_data, aka "risk drivers, that we want to model. It also adjusts
for any latent dates up until the day before Sys.Date.
}
