% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL_stock.R
\name{generate.yf.price.and.returns.datasets}
\alias{generate.yf.price.and.returns.datasets}
\title{Loads Data Sets}
\usage{
generate.yf.price.and.returns.datasets(
  my_sym,
  num_features,
  yearz,
  ticker_file
)
}
\arguments{
\item{my_sym}{a single ticker}

\item{num_features}{number of stocks to add on as features}

\item{yearz}{a look-back in amount of years, can be double or integer}

\item{ticker_file}{a file name within the directory containing a list of tickers}
}
\description{
This function loads the price and returns data sets.
}
