% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_iron_butterfly.R
\name{get.and.price.rolling.options}
\alias{get.and.price.rolling.options}
\title{Calculate Option Stuff}
\usage{
get.and.price.rolling.options(opt_args, price_data, ret_data)
}
\arguments{
\item{opt_args}{an object created from make.opt.args}

\item{price_data}{a data frame with adjusted close data and an index}

\item{ret_data}{a data frame of adj close returns that contains my_sym as
a column}
}
\description{
This function adds in more data that will be helpful in making more accurate
readings on option contract values and their greeks, hopefully...
}
