% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finviz_scraper.R
\name{get_Desired_Tickers_from_Finviz}
\alias{get_Desired_Tickers_from_Finviz}
\title{Scrape Finviz}
\usage{
get_Desired_Tickers_from_Finviz(url)
}
\arguments{
\item{url}{the url to a finviz.com screener page}
}
\description{
This function scrapes and collects data from FinViz.
}
