% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels_wrappers.R
\name{is.DOW.to.STO}
\alias{is.DOW.to.STO}
\title{DOW helper function}
\usage{
is.DOW.to.STO(date, DOW_start)
}
\arguments{
\item{date}{the date we want to test against the supplied DOW}

\item{DOW_start}{the DOW we want each OOS assessment set to start on}
}
\description{
This function returns a boolean that helps the resampling process locate
a user specified DOW.
}
