% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL_stock.R
\name{make.feature.symbols.vec}
\alias{make.feature.symbols.vec}
\title{Reading in tickers for features}
\usage{
make.feature.symbols.vec(file, num_stocks)
}
\arguments{
\item{file}{a file name within the directory containing a list of tickers}

\item{num_stocks}{number of stocks to add on as features}
}
\description{
This function is used to read in a list of tickers. This can be provided by
running get_Desired_Tickers_from_Finviz and saving the second list item as
the file name within the current directory.
}
