% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_funs.R
\name{make.full.future.split}
\alias{make.full.future.split}
\title{Futurize the data}
\usage{
make.full.future.split(data, length_out = 30)
}
\arguments{
\item{data}{a data frame which should be from my_datasets}

\item{length_out}{the investment horizon}
}
\description{
This function take a data frame and extends it out to the investment horizon
}
