% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_funs.R
\name{make.hyperparameter.table}
\alias{make.hyperparameter.table}
\title{Create optimization grid}
\usage{
make.hyperparameter.table(
  initial_balance,
  profit_take_pct,
  stop_loss,
  capital_reserve,
  min_pctl,
  max_pctl,
  pct_below_mid,
  incrament,
  total_param_sets
)
}
\arguments{
\item{initial_balance}{starting account balance}

\item{profit_take_pct}{a sequence}

\item{stop_loss}{a sequence}

\item{capital_reserve}{a sequence}

\item{min_pctl}{minimum strike percentile}

\item{max_pctl}{maximum strike percentile}

\item{pct_below_mid}{threshold below middle percentile strike}

\item{incrament}{the incrament on a parameter sequences}

\item{total_param_sets}{total paramaters to random search}
}
\description{
This function creates a hyperparameter grid that will be random searched
}
