% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_iron_butterfly.R
\name{make.opt.args}
\alias{make.opt.args}
\title{Create option contract arguments}
\usage{
make.opt.args(trimmed_outcomes, oos_resamples, price_data)
}
\arguments{
\item{trimmed_outcomes}{an object created by trim.all.outcomes}

\item{oos_resamples}{the original and unaltered assessment set of resamples}

\item{price_data}{a data frame with adjusted close data and an index}
}
\description{
This function creates a table of option argument. In the future, support for
a continuous dividend yield and risk free rate will be properly canonized.
In the future,
}
