% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels_wrappers.R
\name{make.rolling.custom.resamples}
\alias{make.rolling.custom.resamples}
\title{Customized Resampling}
\usage{
make.rolling.custom.resamples(
  split,
  horizon,
  DOW_start,
  max_slices,
  use_analysis = TRUE
)
}
\arguments{
\item{split}{a split object, preferably a validation split}

\item{horizon}{the investment horizon}

\item{DOW_start}{the DOW we want each OOS assessment set to start on}

\item{max_slices}{the number of folds/slices in our resamples}

\item{use_analysis}{if TRUE, uses the analysis set of the split, if FALSE,
uses the assessment set of the split}
}
\description{
This function makes customized cumulative resamples with a motivation to
keep each OOS set within the resamples starting and ending on the same DOW
}
