% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_iron_butterfly.R
\name{make.shifly.values}
\alias{make.shifly.values}
\title{Make Shifly Values}
\usage{
make.shifly.values(rolling_options)
}
\arguments{
\item{rolling_options}{an object created from get.and.price.rolling.options}
}
\description{
This function generates the rolling shifly values
}
