% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_iron_butterfly.R
\name{make_shifly_grid}
\alias{make_shifly_grid}
\title{Create shifly grid}
\usage{
make_shifly_grid(
  new_iter_tbl,
  min_pctl,
  max_pctl,
  incrament,
  min_hold_days,
  max_hold_days
)
}
\arguments{
\item{new_iter_tbl}{an object created from sim.oos.resamps.and.future.trade}

\item{min_pctl}{minimum strike percentile}

\item{max_pctl}{maximum strike percentile}

\item{incrament}{the increment on a parameter sequences}

\item{min_hold_days}{the minimum number of days to hold the trade}

\item{max_hold_days}{the maximum number of days to hold the trade
which should always be less than the horizon}
}
\description{
This function amends the base grid created from
sim.oos.resamps.and.future.trade
}
