% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels_wrappers.R
\name{plot_best.wf.linear.reg}
\alias{plot_best.wf.linear.reg}
\title{Plot best workflow linear regression}
\usage{
plot_best.wf.linear.reg(tuning_results)
}
\arguments{
\item{tuning_results}{a table created by extract.best.oos.workflow}
}
\description{
Take the output of extract.best.oos.workflow and returns a classic
linear regression plot for observed vs. predicted
}
