% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL_stock.R
\name{remove.young.stocks}
\alias{remove.young.stocks}
\title{Remove stocks that aren't old enough}
\usage{
remove.young.stocks(price_data)
}
\arguments{
\item{price_data}{a data frame with adjusted close data and an index}
}
\description{
This function removes any stocks that do not meet the minimum data
requirement.
}
