% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_iron_butterfly.R
\name{run.shifly.strategy}
\alias{run.shifly.strategy}
\title{Run the Shifly Strategy}
\usage{
run.shifly.strategy(
  out,
  initial_balance,
  capital_reserve,
  profit_take_pct,
  stop_loss,
  hold_days
)
}
\arguments{
\item{out}{an object created from attach.option.returns}

\item{initial_balance}{a single value of the starting account balance}

\item{capital_reserve}{a single value between 0-1}

\item{profit_take_pct}{a single value between 0-1}

\item{stop_loss}{a single value between 0-1}

\item{hold_days}{a single value for a number of days we choose to hold the trade before BTC}
}
\description{
This strategy implements our constraints/parameters on a wealth generation
model centered around rules that govern a short iron butterfly. In the future,
taxes and fees will be accounted for. Random leg assignment will also be
accounted for.
}
