% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL_stock.R
\name{safe.yf.adjusted.price.data.retrieval}
\alias{safe.yf.adjusted.price.data.retrieval}
\title{Download Yahoo Finance data}
\usage{
safe.yf.adjusted.price.data.retrieval(symbol, yearz)
}
\arguments{
\item{symbol}{single/vector of tickers to retrieve adjusted close}

\item{yearz}{a look-back in amount of years, can be double or integer}
}
\description{
This is a safe wrapper to get adjusted close prices from yahoo finance
}
