% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_iron_butterfly.R
\name{shifly_K_grid}
\alias{shifly_K_grid}
\title{Generate shifly grid}
\usage{
shifly_K_grid(min_pctl = 0.05, max_pctl = 0.95, incrament = 0.0025)
}
\arguments{
\item{min_pctl}{minimum strike percentile}

\item{max_pctl}{maximum strike percentile}

\item{incrament}{the increment on a parameter sequences}
}
\description{
Internal: This function samples legs within a range to be tested on the shifly
}
