% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_funs.R
\name{sim.and.backtest.workflow.resamples}
\alias{sim.and.backtest.workflow.resamples}
\title{Fit and Simulate the data sets}
\usage{
sim.and.backtest.workflow.resamples(
  resamples,
  fitt,
  oos_split,
  my_sym,
  price_data,
  num_sims,
  probs
)
}
\arguments{
\item{resamples}{an rset object supplied by make.rolling.custom.resamples}

\item{fitt}{the best workflow embedded in the output of extract.best.oos.workflow}

\item{oos_split}{a split object, preferably NOT the validation set}

\item{my_sym}{a single ticker}

\item{price_data}{a data frame with adjusted close data and an index}

\item{num_sims}{number of runs/calls to MASS::mvrnorm}

\item{probs}{the terminal adj close predicted percentiles to summarise}
}
\description{
Internal -
WARNING: parallelization used. This function is responsible for running
predictions on simulated data. The more simulations, the higher the accuracy,
but accuracy may be marginal after a certain amount. The goal is to return the
percentiles that the model predicts.
}
