% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_funs.R
\name{sim.and.fit}
\alias{sim.and.fit}
\title{Fit and Simulate the data sets}
\usage{
sim.and.fit(
  fitt,
  split,
  my_sym,
  start_price,
  num_sims = 500,
  probs = c(0.05, 0.3, 0.5, 0.7, 0.95)
)
}
\arguments{
\item{fitt}{the best workflow embedded in the output of extract.best.oos.workflow}

\item{split}{a split object, preferably NOT the validation set}

\item{my_sym}{a single ticker}

\item{start_price}{the last price before the prediction occurs}

\item{num_sims}{number of runs/calls to MASS::mvrnorm}

\item{probs}{the terminal adj close predicted percentiles to summarise}
}
\description{
Internal -
WARNING: parallelization used. This function is responsible for running
predictions on simulated data. The more simulations, the higher the accuracy,
but accuracy may be marginal after a certain amount. The goal is to return the
percentiles that the model predicts.
}
