% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_funs.R
\name{sim.my.features}
\alias{sim.my.features}
\title{Simulate Features}
\usage{
sim.my.features(as, an, meanz)
}
\arguments{
\item{as}{the analysis set to take a covariance matrix from}

\item{an}{the assessment set to replace data for}

\item{meanz}{the colmeans of the analysis set}
}
\description{
This function takes the components of a split object and simulates the OOS set
}
