% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_funs.R
\name{sim.oos.resamps.and.future.trade}
\alias{sim.oos.resamps.and.future.trade}
\title{Fit and Simulate}
\usage{
sim.oos.resamps.and.future.trade(
  split,
  tuning_results,
  horizon,
  price_data,
  my_sym,
  num_sims = 1000,
  incrament = 0.0025,
  min_pctl = 0.05,
  max_pctl = 0.95,
  pct_below_mid = 0.1,
  initial_balance = 1e+05,
  profit_take_pct = seq(0.05, 1, 0.01),
  stop_loss = seq(0.05, 1, 0.01),
  capital_reserve = seq(0, 0.2, 0.01),
  total_param_sets = 1000,
  max_asessment_slices = 10
)
}
\arguments{
\item{split}{a split object}

\item{tuning_results}{a table created by extract.best.oos.workflow}

\item{horizon}{the investment horizon}

\item{price_data}{a data frame with adjusted close data and an index}

\item{my_sym}{a single ticker}

\item{num_sims}{number of runs/calls to MASS::mvrnorm}

\item{incrament}{the increment on a parameter sequences}

\item{min_pctl}{minimum strike percentile}

\item{max_pctl}{maximum strike percentile}

\item{pct_below_mid}{threshold below middle percentile strike}

\item{initial_balance}{starting account balance}

\item{profit_take_pct}{a sequence}

\item{stop_loss}{a sequence}

\item{capital_reserve}{a sequence}

\item{total_param_sets}{total paramaters to random search}

\item{max_asessment_slices}{max slices/folds in resamples to test}
}
\description{
This function serves as the simulator for the adjusted close price of the
target asset. Naturally, we can place bounds on the parameters, but watch out!
Different strategies cannot always take the same parameters!
}
