% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_iron_butterfly.R
\name{trim.all.outcomes}
\alias{trim.all.outcomes}
\title{Trimming for the strikes}
\usage{
trim.all.outcomes(all_outcomes, low_leg, mid_leg, upp_leg)
}
\arguments{
\item{all_outcomes}{a list of percentiles and corresponding values for
each OOS fold tested created by sim.oos.resamps.and.future.trade}

\item{low_leg}{a single value for the long put in the strangle}

\item{mid_leg}{a single value for 2 contracts in the short straddle}

\item{upp_leg}{a single value for the long call in the strangle}
}
\description{
This function trims up the simulations into a percentiles table with the
closest strikes we can reasonably place a shifly on. In the future, REAL
historical option values will be used. This is just a cop-out helper function.
}
