% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL_stock.R
\name{yf.load.and.curate.financial.dataset}
\alias{yf.load.and.curate.financial.dataset}
\title{Final formatting for price and return data sets}
\usage{
yf.load.and.curate.financial.dataset(my_sym, num_features, yearz, ticker_file)
}
\arguments{
\item{my_sym}{a single ticker}

\item{num_features}{number of stocks to add on as features}

\item{yearz}{a look-back in amount of years, can be double or integer}

\item{ticker_file}{a file name within the directory containing a list of tickers}
}
\description{
This function loads two data sets: in position 1, the price_data -
"risk drivers", and in position 2, the ret_data - "invariants". We also
format the ret_data to be more easily compatible with the custom ML wrappers.
}
